fix: Email requis sur login feat: Ajout de Traduction feat: Migration diverse feat: Ajout DashboardController feat: Ajout Supplier CRUD feat: Ajout RawMaterial CRUD feat: Mise en forme du design global feat: Template avec des components Twig fix: Divers

+++ src/Twig/Components/NewProductForm.php
@@ -34,7 +34,7 @@ class NewProductForm extends AbstractController
     public string $name = '';
 
     #[LiveProp(writable: true)]
-    public int $price = 0;
+    public float $price = 0;
 
     #[LiveProp(writable: true)]
     #[NotBlank]
